Abm preoductos

+++ src/app/lista-producto/lista-producto.component.ts
@@ -7,10 +7,11 @@ import { ProductoDetalleComponent } from "../producto-detalle/producto-detalle.c
 import { ProductoService } from '../servicios/producto.service'; 
 import { Producto } from '../Clases/producto';
 import { MatIconModule } from '@angular/material/icon';
-import { FiltroProductoPipe } from '../pipes/filtro-producto.pipe';
+import { FiltroProductosPipe } from '../pipes/filtro-productos.pipe'; 
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 @Component({
   selector: 'app-lista-producto',
-  imports: [MatListModule, CommonModule, MatCardModule, MatIconModule, ProductoDetalleComponent,FiltroProductoPipe],
+  imports: [MatListModule, CommonModule, MatCardModule, MatIconModule, FormsModule,ProductoDetalleComponent,FiltroProductosPipe,ReactiveFormsModule],
   templateUrl: './lista-producto.component.html',
   styleUrl: './lista-producto.component.css'
 })
@@ -20,8 +21,14 @@ export class ListaProductoComponent {
 
   constructor(private productoService: ProductoService, private router: Router) {}
 
-  ngOnInit() {
-    this.productoService.obtenerProductos().subscribe(data => this.productos = data);
+  ngOnInit(): void {
+    this.cargarProductos();
+  }
+
+  cargarProductos() {
+    this.productoService.obtenerProductos().subscribe(data => {
+      this.productos = data;
+    });
   }
 
   editar(id: number) {
@@ -29,9 +36,11 @@ export class ListaProductoComponent {
   }
 
   eliminar(id: number) {
-    this.productoService.eliminarProducto(id).subscribe(() => {
-      this.productos = this.productos.filter(p => p.id !== id);
-    });
+    if (confirm('¿Estás segura/o de eliminar este producto?')) {
+      this.productoService.eliminarProducto(id).subscribe(() => {
+        this.productos = this.productos.filter(p => p.id !== id);
+      });
+    }
   }
 
   agregarProducto() {
